# master graph generation & parallelism