--- a/siteupdate/cplusplus/functions/sql_file.cpp
+++ b/siteupdate/cplusplus/functions/sql_file.cpp
@@ -351,7 +351,9 @@ void sqlfile1
 	}
 	sqlfile << ";\n";
 	sqlfile.close();
+      #ifdef threading_enabled
 	std::cout << '\n' + et->et() + "Pause writing database file " + args->databasename + ".sql.\n" << std::flush;
+      #endif
      }
 
 void sqlfile2