Addresses most of the speed side of #200.
* **C++:** Master graph generation time 11.6s -> 8.5s, single-threaded, a 36% speedup. Multithreaded, master graphs are written alongside subgaphs, saving us another ~3.2s with 4 threads on BiggaTomato.
* **Python:** Master graph generation time 29.7s -> 26s, a 14% speedup. Not as good as I might have hoped for but an improvement nonetheless.